net/faraday: Use ncsi_stop_dev()
We're using ncsi_suspend_dev() to tear down NCSI device when bringing
down the network interface. In ncsi_suspend_dev(), NCSI packets might
be sent over NCSI link to stop the interface physically. However, the
NCSI packets might not be sent successfully when the NCSI interface
is down. In that case, we potentially release the buffers and their
ring for twice exceptionally, which eventually leads to kernel crash.

To resolve the above problem, this calls ncsi_stop_dev() to tear down
the NCSI interface without sending NCSI packets over the link. At the
same time, a flag (priv->enabled) introduced to avoid releasing the
buffers and their rings for twice.

Signed-off-by: Gavin Shan <gwshan@linux.vnet.ibm.com>
Signed-off-by: Joel Stanley <joel@jms.id.au>